Move grouplayout to src root
